feat: use floats (0-1.0) to store actions instead of ints (0-3) (#20)

+++ agent/models/experience.py
@@ -1,9 +1,9 @@
 from pydantic import BaseModel
 
 class Experience(BaseModel):
-    state: list[int]
-    action: int 
+    state: list[float]
+    action: int
     reward: float
-    next_state: list[int]
+    next_state: list[float]
     done: bool
